Rename Kite Commands file
